fix directory task order, seperaate queue tasks into a file
